Bot-4/actions/requirements-actions.txt
newspaper3k==0.2.8
pymongo==3.10.1
nltk==3.6.5